LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-async-fetch-mutation-remove-middle-expected.txt


FAIL Removing middle placeholder before fetch excludes it from final adoptedStyleSheets. assert_equals: Before mutation: expected 3 placeholder(s). expected 3 but got 0
